refactor models/fhir to respective model classes.
