+++ portal/views/assessment_engine.py
@@ -30,15 +30,15 @@ from ..models.assessment_status import (
     overall_assessment_status,
 )
 from ..models.client import validate_origin
-from ..models.fhir import (
-    EC,
+from ..models.encounter import EC
+from ..models.intervention import INTERVENTION
+from ..models.questionnaire import Questionnaire
+from ..models.questionnaire_bank import QuestionnaireBank
+from ..models.questionnaire_response import (
     QuestionnaireResponse,
     aggregate_responses,
     generate_qnr_csv,
 )
-from ..models.intervention import INTERVENTION
-from ..models.questionnaire import Questionnaire
-from ..models.questionnaire_bank import QuestionnaireBank
 from ..models.role import ROLE
 from ..models.user import User, current_user, get_user_or_abort
 from ..trace import dump_trace, establish_trace
