+++ tests/test_fhir.py
@@ -9,11 +9,9 @@ import pytz
 from portal.extensions import db
 from portal.models.codeable_concept import CodeableConcept
 from portal.models.coding import Coding
-from portal.models.fhir import (
-    FHIR_datetime,
-    QuestionnaireResponse,
-    ValueQuantity,
-)
+from portal.date_tools import FHIR_datetime
+from portal.models.questionnaire_response import QuestionnaireResponse
+from portal.models.value_quantity import ValueQuantity
 from portal.system_uri import SNOMED
 from tests import TEST_USER_ID, TestCase
